Better organization and some documentation
