+++ apps/shadcn-example/src/app/layout.tsx
@@ -1,10 +1,10 @@
+import "./globals.css";
 import type { Metadata } from "next";
 import { Inter as FontSans } from "next/font/google";
-import "./globals.css";
-import { cn } from "@/lib/utils";
 import { ThemeProvider } from "@/components/theme-provider";
-import { WalletProvider } from "@/components/wallet-connection";
 import { Toaster } from "@/components/ui/toaster";
+import { WalletProvider } from "@/components/wallet-provider";
+import { cn } from "@/lib/utils";
 
 const fontSans = FontSans({
   subsets: ["latin"],
