--- apps/shadcn-example/src/components/wallet-connection/index.tsx
@@ -1,113 +1,0 @@
-"use client";
-
-import { Network } from "@aptos-labs/ts-sdk";
-import {
-  AptosWalletAdapterProvider,
-  truncateAddress,
-  useWallet,
-} from "@aptos-labs/wallet-adapter-react";
-import { BitgetWallet } from "@bitget-wallet/aptos-wallet-adapter";
-import { IdentityConnectWallet } from "@identity-connect/wallet-adapter-plugin";
-import { MartianWallet } from "@martianwallet/aptos-wallet-adapter";
-import { MSafeWalletAdapter } from "@msafe/aptos-wallet-adapter";
-import { OKXWallet } from "@okwallet/aptos-wallet-adapter";
-import { PontemWallet } from "@pontem/wallet-adapter-plugin";
-import { TrustWallet } from "@trustwallet/aptos-wallet-adapter";
-import { FewchaWallet } from "fewcha-plugin-wallet-adapter";
-import { Copy, LogOut } from "lucide-react";
-import { PetraWallet } from "petra-plugin-wallet-adapter";
-import { PropsWithChildren, useCallback, useState } from "react";
-import { Button } from "../ui/button";
-import { Dialog, DialogTrigger } from "../ui/dialog";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "../ui/dropdown-menu";
-import { useToast } from "../ui/use-toast";
-import { ConnectWalletDialog } from "./connect-wallet";
-
-export const WalletProvider = ({ children }: PropsWithChildren) => {
-  const { toast } = useToast();
-
-  const wallets = [
-    new IdentityConnectWallet("57fa42a9-29c6-4f1e-939c-4eefa36d9ff5", {
-      networkName: Network.TESTNET,
-    }),
-    new BitgetWallet(),
-    new FewchaWallet(),
-    new MartianWallet(),
-    new MSafeWalletAdapter(),
-    new PontemWallet(),
-    new TrustWallet(),
-    new OKXWallet(),
-    new PetraWallet(),
-  ];
-
-  return (
-    <AptosWalletAdapterProvider
-      plugins={wallets}
-      autoConnect={true}
-      onError={(error) => {
-        toast({
-          variant: "destructive",
-          title: "Error",
-          description: error || "Unknown wallet error",
-        });
-      }}
-    >
-      {children}
-    </AptosWalletAdapterProvider>
-  );
-};
-
-export function WalletConnectionButton() {
-  const { disconnect, account, connected } = useWallet();
-  const { toast } = useToast();
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-
-  const closeDialog = useCallback(() => setIsDialogOpen(false), []);
-
-  const copyAddress = useCallback(async () => {
-    if (!account?.address) return;
-    try {
-      await navigator.clipboard.writeText(account.address);
-      toast({
-        title: "Success",
-        description: "Copied wallet address to clipboard.",
-      });
-    } catch {
-      toast({
-        variant: "destructive",
-        title: "Error",
-        description: "Failed to copy wallet address.",
-      });
-    }
-  }, [account?.address, toast]);
-
-  return connected ? (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button>
-          {account?.ansName || truncateAddress(account?.address) || "Unknown"}
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end">
-        <DropdownMenuItem onSelect={copyAddress} className="gap-2">
-          <Copy className="h-4 w-4" /> Copy address
-        </DropdownMenuItem>
-        <DropdownMenuItem onSelect={disconnect} className="gap-2">
-          <LogOut className="h-4 w-4" /> Disconnect
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  ) : (
-    <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
-      <DialogTrigger asChild>
-        <Button>Connect a Wallet</Button>
-      </DialogTrigger>
-      <ConnectWalletDialog close={closeDialog} />
-    </Dialog>
-  );
-}
