+++ apps/shadcn-example/src/components/wallet-selector.tsx
@@ -1,0 +1,145 @@
+"use client";
+
+import {
+  AptosWallet,
+  WalletItem,
+  isInstallRequired,
+  partitionWallets,
+  truncateAddress,
+  useWallet,
+} from "@aptos-labs/wallet-adapter-react";
+import { ChevronDown, Copy, LogOut } from "lucide-react";
+import { useCallback, useState } from "react";
+import { Button } from "./ui/button";
+import {
+  Collapsible,
+  CollapsibleContent,
+  CollapsibleTrigger,
+} from "./ui/collapsible";
+import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "./ui/dialog";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "./ui/dropdown-menu";
+import { useToast } from "./ui/use-toast";
+
+export function WalletSelector() {
+  const { disconnect, account, connected } = useWallet();
+  const { toast } = useToast();
+  const [isDialogOpen, setIsDialogOpen] = useState(false);
+
+  const closeDialog = useCallback(() => setIsDialogOpen(false), []);
+
+  const copyAddress = useCallback(async () => {
+    if (!account?.address) return;
+    try {
+      await navigator.clipboard.writeText(account.address);
+      toast({
+        title: "Success",
+        description: "Copied wallet address to clipboard.",
+      });
+    } catch {
+      toast({
+        variant: "destructive",
+        title: "Error",
+        description: "Failed to copy wallet address.",
+      });
+    }
+  }, [account?.address, toast]);
+
+  return connected ? (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button>
+          {account?.ansName || truncateAddress(account?.address) || "Unknown"}
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent align="end">
+        <DropdownMenuItem onSelect={copyAddress} className="gap-2">
+          <Copy className="h-4 w-4" /> Copy address
+        </DropdownMenuItem>
+        <DropdownMenuItem onSelect={disconnect} className="gap-2">
+          <LogOut className="h-4 w-4" /> Disconnect
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  ) : (
+    <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
+      <DialogTrigger asChild>
+        <Button>Connect a Wallet</Button>
+      </DialogTrigger>
+      <ConnectWalletDialog close={closeDialog} />
+    </Dialog>
+  );
+}
+
+interface ConnectWalletDialogProps {
+  close: () => void;
+}
+
+function ConnectWalletDialog({ close }: ConnectWalletDialogProps) {
+  const { wallets } = useWallet();
+  const { defaultWallets, moreWallets } = partitionWallets(wallets ?? []);
+
+  return (
+    <DialogContent className="max-h-screen overflow-auto">
+      <DialogHeader>
+        <DialogTitle>Connect Wallet</DialogTitle>
+      </DialogHeader>
+      <div className="flex flex-col gap-3 pt-3">
+        {defaultWallets.map((wallet) => (
+          <WalletRow key={wallet.name} wallet={wallet} onConnect={close} />
+        ))}
+      </div>
+      <Collapsible className="flex flex-col gap-4">
+        <CollapsibleTrigger asChild>
+          <Button size="sm" variant="ghost" className="gap-2">
+            More wallets <ChevronDown />
+          </Button>
+        </CollapsibleTrigger>
+        <CollapsibleContent className="flex flex-col gap-3">
+          {moreWallets.map((wallet) => (
+            <WalletRow key={wallet.name} wallet={wallet} onConnect={close} />
+          ))}
+        </CollapsibleContent>
+      </Collapsible>
+    </DialogContent>
+  );
+}
+
+interface WalletRowProps {
+  wallet: AptosWallet;
+  onConnect?: () => void;
+}
+
+function WalletRow({ wallet, onConnect }: WalletRowProps) {
+  return (
+    <WalletItem
+      wallet={wallet}
+      onConnect={onConnect}
+      className="flex items-center justify-between px-4 py-3 gap-4 border rounded-md"
+    >
+      <div className="flex items-center gap-4">
+        <WalletItem.Icon className="h-6 w-6" />
+        <WalletItem.Name className="text-base font-normal" />
+      </div>
+      {isInstallRequired(wallet) ? (
+        <Button size="sm" variant="ghost" asChild>
+          <WalletItem.InstallLink />
+        </Button>
+      ) : (
+        <WalletItem.ConnectButton asChild>
+          <Button size="sm">Connect</Button>
+        </WalletItem.ConnectButton>
+      )}
+    </WalletItem>
+  );
+}
